Time: 0 ms (100.00%) | Memory: 71.8 MB (51.54%) - LeetSync

+++ 53-maximum-subarray/maximum-subarray.cpp
@@ -7,7 +7,7 @@ class Solution {
 public:
     int maxSubArray(vector<int>& nums) {
         int n = nums.size();
-        long long sum = 0 , max_sum = INT_MIN;
+        int sum = 0 , max_sum = INT_MIN;
 
         for(int i = 0 ; i < n ; ++i){
             sum += nums[i];
